back/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.1
pytest==7.4.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
